# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: ShopPro 智能SCRM系统 API
  version: 1.0.0
  description: ShopPro 智能客户关系管理系统 RESTful API 规范 - 提供完整的客户管理、销售跟进、数据分析等功能接口
servers:
  - url: http://localhost:8080/api
    description: 本地开发
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageNo:
      name: pageNo
      in: query
      schema: { type: integer, minimum: 1, default: 1, example: 1 }
      description: 页码，最小为 1，默认 1
    PageSize:
      name: pageSize
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 10, example: 10 }
      description: 每页大小，范围 [1,100]，默认 10
    Sort:
      name: sort
      in: query
      schema: { type: string, example: "createdAt,desc" }
    Keyword:
      name: keyword
      in: query
      schema: { type: string }
  schemas:
    ApiResponse:
      type: object
      properties:
        code: { type: integer }
        message: { type: string }
        data: {}
...
    PageProductResult:
      type: object
      properties:
        records:
          type: array
          items: { $ref: '#/components/schemas/Product' }
        current: { type: integer }
        size: { type: integer }
        total: { type: integer }
        pages: { type: integer }
...
    Product:
      type: object
      required: [productName, price]
      properties:
        id: { type: integer, format: int64 }
        sku: { type: string }
        productName: { type: string }
        categoryId: { type: integer, format: int64, nullable: true }
        price: { type: number, format: double }
        stock: { type: integer }
        status: { type: string, enum: [active, inactive], default: active }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
paths:
...
  /products:
    get:
      tags: [Products]
      summary: 分页查询产品
      parameters:
        - $ref: '#/components/parameters/PageNo'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/Sort'
        - $ref: '#/components/parameters/Keyword'
        - name: status
          in: query
          schema: { type: string }
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageProductResult'
    post:
      tags: [Products]
      summary: 创建产品
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Product' }
      responses:
        '201': { description: Created }
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
  /products/{id}:
    get:
      tags: [Products]
      summary: 获取产品详情
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200': { description: OK }
        '404':
          description: 未找到产品
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
        '500':
          description: 服务器内部错误
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
    put:
      tags: [Products]
      summary: 更新产品
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Product' }
      responses:
        '200': { description: OK }
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
    delete:
      tags: [Products]
      summary: 删除产品
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '204': { description: No Content }
...